Move array-related tests to their own module
